Clean up Prometheus-based measurements unit testing framework
