+++ clusterloader2/pkg/measurement/common/slos/api_responsiveness_prometheus_test.go
@@ -21,15 +21,12 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"io/ioutil"
-	"os"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/assert"
-	"gopkg.in/yaml.v2"
 	"k8s.io/klog"
 	"k8s.io/perf-tests/clusterloader2/pkg/errors"
 	"k8s.io/perf-tests/clusterloader2/pkg/measurement"
@@ -89,55 +86,6 @@ type summaryEntry struct {
 
 type fakeQueryExecutor struct {
 	samples []*sample
-}
-
-type rule struct {
-	Expr   string `yaml:"expr"`
-	Record string `yaml:"record"`
-	Labels struct {
-		Quantile string `yaml:"quantile"`
-	} `yaml:"labels"`
-}
-
-type group struct {
-	Name  string `yaml:"name"`
-	Rules []rule `yaml:"rules"`
-}
-
-//prometheusRuleManifest mimics the structure of PrometheusRule object used by prometheus operator
-//https://github.com/prometheus-operator/prometheus-operator/blob/main/pkg/apis/monitoring/v1/types.go#L1393
-type prometheusRuleManifest struct {
-	Spec struct {
-		Groups []group `yaml:"groups"`
-	} `yaml:"spec"`
-}
-
-func createRulesFile(rulesManifestFile string) (*os.File, error) {
-	r, err := ioutil.ReadFile(rulesManifestFile)
-	if err != nil {
-		return nil, err
-	}
-
-	rulesManifest := new(prometheusRuleManifest)
-	err = yaml.Unmarshal(r, rulesManifest)
-	if err != nil {
-		return nil, err
-	}
-
-	tempFile, err := ioutil.TempFile("", "")
-	if err != nil {
-		return nil, err
-	}
-	b, err := yaml.Marshal(rulesManifest.Spec)
-	if err != nil {
-		return nil, err
-	}
-
-	_, err = tempFile.Write(b)
-	if err != nil {
-		return nil, err
-	}
-	return tempFile, nil
 }
 
 func (ex *fakeQueryExecutor) Query(query string, queryTime time.Time) ([]*model.Sample, error) {
@@ -268,14 +216,9 @@ func TestAPIResponsivenessSLOFailures(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
-			f, err := createRulesFile("../../../prometheus/manifests/prometheus-rules.yaml")
+			executor, err := executors.NewPromqlExecutor(fmt.Sprintf("../testdata/api_responsiveness_prometheus/%s", tc.testSeriesFile))
 			if err != nil {
-				t.Fatalf("failed to create ruels file: %v", err)
-			}
-			defer os.Remove(f.Name())
-			executor, err := executors.NewPromqlExecutor(fmt.Sprintf("testdata/%s", tc.testSeriesFile), f.Name())
-			if err != nil {
-				t.Fatalf("failed to create ruels file: %v", err)
+				t.Fatalf("failed to create PromQL executor: %v", err)
 			}
 			defer executor.Close()
 			gatherer := &apiResponsivenessGatherer{}
