+++ clusterloader2/pkg/measurement/common/testdata/api_responsiveness_prometheus/slo_pass.yaml
