+++ clusterloader2/pkg/measurement/common/testdata/cilium_endpoint_propagation_delay/default_slo_pass.yaml
