+++ clusterloader2/pkg/measurement/common/testdata/container_restarts/double_restart_of_apiserver.yaml
