+++ clusterloader2/pkg/measurement/common/testdata/container_restarts/no_restarts.yaml
