+++ clusterloader2/pkg/measurement/common/testdata/container_restarts/two_apiserver_replicas_restarts.yaml
